+++ src/Storage/Storage.Management/File/UpdateAzureStorageShare.cs
@@ -13,6 +13,7 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Commands.Management.Storage.Models;
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 using Microsoft.Azure.Management.Storage;
 using Microsoft.Azure.Management.Storage.Models;
@@ -126,6 +127,28 @@ namespace Microsoft.Azure.Commands.Management.Storage
             IgnoreCase = true)]
         public string RootSquash { get; set; }
 
+        [Parameter(
+           Mandatory = false,
+           HelpMessage = "Access tier for specific share. StorageV2 account can choose between TransactionOptimized (default), Hot, and Cool. FileStorage account can choose Premium.")]
+        [ValidateSet(ShareAccessTier.TransactionOptimized,
+            ShareAccessTier.Premium,
+            ShareAccessTier.Hot,
+            ShareAccessTier.Cool,
+           IgnoreCase = true)]
+        [ValidateNotNullOrEmpty]
+        public string AccessTier
+        {
+            get
+            {
+                return accessTier;
+            }
+            set
+            {
+                accessTier = value;
+            }
+        }
+        private string accessTier = null;
+        
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -162,7 +185,8 @@ namespace Microsoft.Azure.Commands.Management.Storage
                                     new FileShare(
                                         metadata: MetadataDictionary,
                                         shareQuota: shareQuota,
-                                        rootSquash: this.RootSquash));
+                                        rootSquash: this.RootSquash,
+                                        accessTier: accessTier));
 
                 WriteObject(new PSShare(Share));
             }