+++ src/Storage/Storage.Management/Models/PSShare.cs
@@ -39,8 +39,6 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.Etag = share.Etag;
             this.LastModifiedTime = share.LastModifiedTime;
             this.QuotaGiB = share.ShareQuota;
-            this.EnabledProtocols = share.EnabledProtocols;
-            this.RootSquash = share.RootSquash;
             this.Version = share.Version;
             this.Deleted = share.Deleted;
             this.DeletedTime = share.DeletedTime;
@@ -64,8 +62,6 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.Etag = share.Etag;
             this.LastModifiedTime = share.LastModifiedTime;
             this.QuotaGiB = share.ShareQuota;
-            this.EnabledProtocols = share.EnabledProtocols;
-            this.RootSquash = share.RootSquash;
             this.Version = share.Version;
             this.Deleted = share.Deleted;
             this.DeletedTime = share.DeletedTime;